+++ src/Models/Site.php
@@ -332,12 +332,15 @@ class Site extends TerminusModel implements ConfigAwareInterface, ContainerAware
     {
         $settings = $this->get('settings');
 
-        $created_date = is_numeric($created = $this->get('created')) ? $created : strtotime($created);
+        $date_format = $this->getConfig()->get('date_format');
+        $format_date = function ($attribute) use ($date_format) {
+            return date($date_format, is_numeric($date = $this->get($attribute)) ? $date : strtotime($date));
+        };
         $data = [
             'id' => $this->id,
             'name' => $this->get('name'),
             'label' => $this->get('label'),
-            'created' => date($this->getConfig()->get('date_format'), $created_date),
+            'created' => $format_date('created'),
             'framework' => $this->get('framework'),
             'organization' => $this->get('organization'),
             'service_level' => $this->get('service_level'),
@@ -348,6 +351,7 @@ class Site extends TerminusModel implements ConfigAwareInterface, ContainerAware
             'holder_id' => $this->get('holder_id'),
             'owner' => $this->get('owner'),
             'frozen' => $this->isFrozen() ? 'true' : 'false',
+            'last_frozen_at' => $format_date('last_frozen_at'),
         ];
         if (isset($this->tags)) {
             $data['tags'] = implode(',', $this->tags->ids());